Detailed information about the movie [DONE] Need to create results route
