+++ package.json
@@ -3,16 +3,28 @@
   "version": "1.0.0",
   "main": "server.js",
   "scripts": {
-    "start": "node server.js",
+    "start": "nodemon --exec node index.js -e .js,.ejs,.json",
     "lint:js": "node_modules/eslint/bin/eslint.js ./ ./**/*.js --fix; exit 0",
     "lint:css": "node_modules/csslint/cli.js public/css/; exit 0"
   },
   "license": "ISC",
   "dependencies": {
+    "axios": "^0.26.0",
     "dotenv": "^8.1.0",
     "ejs": "^3.0.1",
     "express": "^4.17.1",
     "express-ejs-layouts": "^2.5.0",
     "morgan": "^1.9.1"
-  }
+  },
+  "description": "We'll be creating an app that connects to [OMDB](http://www.omdbapi.com), a public movie API. You will need a key. Go to the site to register for a free one. Keep API keys out of public repos!",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/br1anlee/express-apis-omdb.git"
+  },
+  "keywords": [],
+  "author": "",
+  "bugs": {
+    "url": "https://github.com/br1anlee/express-apis-omdb/issues"
+  },
+  "homepage": "https://github.com/br1anlee/express-apis-omdb#readme"
 }
