+++ views/detail.ejs
@@ -1,2 +1,7 @@
-<% // Render individual movie details here %>
-<% // This is how to do comments in EJS, by the way... %>
+<h2><%= result.Title %></h2>
+<img src="<%= result.Poster %>" alt="a picture of the movie">
+<p><b><%= result.imdbID %></b></p>
+<p><%= result.Genre %></p>
+<p><%= result.Release %></p>
+<p><%= result.Director %></p>
+<p><%= result.Plot %></p>
