+++ views/index.ejs
@@ -1,2 +1,10 @@
 <h1>OMDB Movie Search</h1>
 <p>Search for a movie below...</p>
+<!-- GET method is required [client is sending data]-->
+<form method="GET" action="/omdb">
+    <label for="omdb-search">Search</label>
+    <!-- id must be = to the for label -->
+    <input type="text" name="search" placeholder="enter your search" id="omdb-search">
+    <!-- submit button -->
+    <input type="submit">
+</form>
