+++ views/layout.ejs
@@ -4,11 +4,11 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <link rel="stylesheet" href="">
   <title>OMDB Movie Search</title>
 </head>
 <body>
   <nav>
-    <a href="/">home</a>
   </nav>
   <main>
     <%- body %>
